refactor: 각 코드를 도메인 계층으로 리팩토링
